Add a second formatter module for logger
